+++ tasks.py
@@ -56,6 +56,5 @@ def tree(c: Context):
 @task
 def update(c: Context):
     """Update dependencies"""
-    # c.run("poetry self update", echo=True, pty=True)
     c.run("poetry up --latest --only=dev", echo=True, pty=True)
     c.run("poetry show --outdated --why --with=dev", echo=True, pty=True)