+++ CHANGELOG.md
@@ -5,8 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 <br>
 
 ## Unreleased
-* Change `aiohttp` requirement to `^3.9` (from `~3.9.0b0`)
-* Change `pydantic` requirement to `^2.4` (from `~2.4`)
+* Relax `aiohttp` requirement to `^3.9` (from `~3.9.0b0`)
+* Relax `pydantic` requirement to `^2.4` (from `~2.4`)
 * Add `py.typed` to make the package PEP 561 compatible
 
 <br>