+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 <br>
 
-## Unreleased
+## [0.3.0] – 2024-04-27
 * Implement `/api/4/tags/popular` endpoint
 * Relax `aiohttp` requirement to `^3.9` (from `~3.9.0b0`)
 * Relax `pydantic` requirement to `^2.4` (from `~2.4`)
@@ -31,3 +31,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 [0.1.0]: https://github.com/timwie/aio-taginfo/releases/tag/v0.1.0
 [0.2.0]: https://github.com/timwie/aio-taginfo/releases/tag/v0.2.0
+[0.3.0]: https://github.com/timwie/aio-taginfo/releases/tag/v0.3.0